Clean Search classes
